requirements.txt
mysql-connector-python
pytz
paramiko